adds middlewares

+++ src/Services/Shared/Groceteria.Shared/Core/Pagination.cs
@@ -2,7 +2,7 @@
 {
     public class Pagination<T> where T: class
     {
-        public Pagination(int pageIndex, int pageSize, int count, IReadOnlyList<T> data)
+        public Pagination(int pageIndex, int pageSize, long count, IReadOnlyList<T> data)
         {
             PageIndex = pageIndex;
             PageSize = pageSize;
@@ -12,7 +12,7 @@
 
         public int PageIndex { get; set; }
         public int PageSize { get; set; }
-        public int Count { get; set; }
+        public long Count { get; set; }
         public IReadOnlyList<T> Data { get; set; }
     }
 }
